Improved the find method
We now nitialize the objects with
just one db query. This should improve the performance a lot!
It can however result in not all properties being set
(when using gridfs or later also a projection when I add this).
Therefore, I also improved the update_from_db method
to have a new functionality so that it can update a list of
selected properties.